pyproject.toml
[project]
name = "researcher"
version = "0.1.0"
description = "A web research agent built with the Agent-to-Agent (A2A) protocol that performs comprehensive web searches using DuckDuckGo and returns detailed results with cited sources."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
     "ai-agent2agent>=0.1.0",
     "crewai>=0.28.0",
     "langchain>=0.1.0",
     "langchain-community>=0.0.10",
     "pydantic>=2.0.0",
     "python-dotenv>=1.0.0",
     "click>=8.0.0",
]
